requirements.txt
moviepy
whisper
google-generativeai
openpyxl
python-dotenv
pytz